LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/timeline-offset-keyframes-hidden-subject-expected.txt


FAIL Keyframes with timeline-offsets ignored when timeline is inactive assert_equals: Should have one animation attached to the view-timeline expected 1 but got 0
